Move OwnerSet to its own file
